// secure-agents/frontend/src/client/types.gen.ts
export type FeedbackPublic = {
  /**
   * Feedback Type
   */
  feedback_type: string;
  /**
   * Feedback Comment
   */
  feedback_comment?: string | null;
  /**
   * Rated Message
   */
  rated_message?: string | null;
  /**
   * User Message Before
   */
  user_message_before?: string | null;
  /**
   * Id
   */
  id: string;
};

/**
 * FeedbackUpdate
 */
export type FeedbackUpdate = {
  /**
   * Feedback Type
   */
  feedback_type?: string | null;
  /**
   * Feedback Comment
   */
  feedback_comment?: string | null;
  /**
   * Rated Message
   */
  rated_message?: string | null;
  /**
   * User Message Before
   */
  user_message_before?: string | null;
};